Support multi-region path computation in multidomain networks. This requires collaboration between peer narbs and replay of MRN path computation procedure.
git-svn-id: file:///var/www/svn/narb-sw/trunk@34 aacf9c1e-ac80-435b-8e83-6affbfa3956d

+++ dragon-sw-plus/narb/narb_config.hh
@@ -68,8 +68,9 @@ enum narb_working_mode {
 // LSP computation options
 #define LSP_OPT_STRICT  ((u_int32_t)(0x01 << 16)) //otherwise LSP_OPT_LOOSE
 #define LSP_OPT_PREFERRED ((u_int32_t)(0x02 << 16)) //otherwise LSP_OPT_ONLY
-#define LSP_OPT_MRN 0x04 << 16  //otherwise SINGLE_REGION
-#define LSP_OPT_BIDIRECTIONAL 0x10 << 16  // otherwise UNIDIRECTIONAL
+#define LSP_OPT_MRN ((u_int32_t)0x04 << 16)  //otherwise SINGLE_REGION
+//#define LSP_OPT_MRN_RELAY ((u_int32_t)0x08 << 16)  //relay path computation for multiregion multidomain networks
+#define LSP_OPT_BIDIRECTIONAL ((u_int32_t)0x10 << 16)  // otherwise UNIDIRECTIONAL
 #define LSP_OPT_E2E_VTAG  ((u_int32_t)(0x20 << 16)) //otherwise Untgged VLAN for E2E Ethernet
 
 class SystemConfig
